clean-code исправил код на более удобно читаемый, добавил комментарии, исправил ошибки
