--- lesson3/clean_code/badCode2/bad-code/src/main/java/com/bad/code2/Shape3D.java
@@ -1,8 +1,0 @@
-package com.bad.code2;
-
-public interface Shape3D {
-    Double getX();
-    Double getY();
-    Double getZ();
-    Double getVolume();
-}
