+++ lesson3/clean_code/badCode2/bad-code/src/main/java/com/bad/code2/shapeInterfaces/Shape2D.java
@@ -1,0 +1,11 @@
+package main.java.com.bad.code2.shapeInterfaces;
+
+/*
+Интерфейс с методом для расчёта периметра 2D фигур
+ */
+public interface Shape2D {
+    /*
+    Метод для расчёта периметра фигуры
+     */
+    Double getPerimeter();
+}
