+++ lesson3/clean_code/badCode2/bad-code/src/main/java/com/bad/code2/shapeInterfaces/Shape3D.java
@@ -1,0 +1,11 @@
+package main.java.com.bad.code2.shapeInterfaces;
+
+/*
+Интерфейс с методом для расчёта объёма 3D фигур
+ */
+public interface Shape3D {
+    /*
+    Метод для расчёта объёма фигуры
+     */
+    Double getVolume();
+}
