+++ lesson3/clean_code/badCode2/bad-code/src/main/java/com/bad/code2/shapes/Qube.java
@@ -1,9 +1,19 @@
-package com.bad.code2;
+package main.java.com.bad.code2.shapes;
+
+import main.java.com.bad.code2.shapeInterfaces.Shape3D;
 
 public class Qube implements Shape3D {
+
+    /*
+    Переменные для хранения координат для отображения фигуры
+     */
     private Double x;
     private Double y;
     private Double z;
+
+    /*
+    Переменная, которая хранит значение стороны куба
+     */
     private Double edgeSize;
 
     public Qube(Double centerX, Double centerY, Double centerZ, Double s) {
@@ -13,21 +23,21 @@ public class Qube implements Shape3D {
         this.edgeSize = s;
     }
 
-    @Override
     public Double getX() {
         return x;
     }
 
-    @Override
     public Double getY() {
         return y;
     }
 
-    @Override
     public Double getZ() {
         return z;
     }
 
+    /*
+    Метод, который расчитывает объём фигуры и возвращает его.
+     */
     @Override
     public Double getVolume() {
         return Math.pow(edgeSize, 3);
